backend/requirements.txt
# Tsifulator.ai — Python Dependencies
# Install with: pip install -r requirements.txt

# Web server (the brain's API layer)
fastapi==0.115.0
uvicorn==0.30.6

# Anthropic (Claude — the AI engine)
anthropic==0.34.0

# Database + Vector Memory (Supabase)
supabase==2.7.2

# Environment variable management
python-dotenv==1.0.1

# Excel file manipulation (for local .xlsx files)
openpyxl==3.1.5

# Utilities
httpx==0.27.2
pydantic==2.8.2
